[while] 100까지의 수를 입력 받아 합, 평균, 입력 갯수 출력
0보다 작은 수를 입력하면 프로그램 종료
100보다 큰 수는 합, 평균, 입력 갯수에서 제외

+++ Test_16.java
@@ -24,7 +24,7 @@ public class Test_16 {
 			
 			cnt++;
 			sum += n;
-			average = (sum / cnt);
+			average = ((double)sum / (double)cnt);
 		}
 		System.out.println("합 : " + sum);
 		System.out.println("평균 : " + average);
